Assignment 2 Submission
- Ready for A2 submission.
- Added 2 weapon upgrades: Sniper and Machine Gun.
- Added experience/level meter UI and tweaked leveling system.
- AI Drones now pursue objects if no differently teamed Drone is in range.
- AI Drones spawn automatically.

+++ Drone Strikers/Assets/Scripts/Combat/CombatInterfaces.cs
@@ -5,7 +5,7 @@
         /// <summary>
         ///     The amount of damage this source deals on contact.
         /// </summary>
-        int ContactDamage { get; }
+        float ContactDamage { get; }
 
         /// <summary>
         ///     The destruction context receiver of the instigator of this damage source, if any.
