+++ Drone Strikers/Assets/Scripts/Combat/DamageContext.cs
@@ -4,7 +4,7 @@ namespace DroneStrikers.Combat
 {
     public struct DamageContext
     {
-        public int DamageAmount { get; }
+        public float DamageAmount { get; }
         public GameObject SourceObject { get; }
         public Team SourceTeam { get; }
         public IDestructionContextReceiver InstigatorContextReceiver { get; }
@@ -16,7 +16,7 @@ namespace DroneStrikers.Combat
         /// <param name="sourceObject"> The GameObject that is the source of the damage. </param>
         /// <param name="sourceTeam"> The team of the source GameObject. </param>
         /// <param name="instigatorContextReceiver"> The IDestructionContextReceiver of the attack's instigator, if any. </param>
-        public DamageContext(int damageAmount, GameObject sourceObject, Team sourceTeam, IDestructionContextReceiver instigatorContextReceiver)
+        public DamageContext(float damageAmount, GameObject sourceObject, Team sourceTeam, IDestructionContextReceiver instigatorContextReceiver)
         {
             DamageAmount = damageAmount;
             SourceObject = sourceObject;
