+++ Drone Strikers/Assets/Scripts/Combat/SimpleBodyDamageSource.cs
@@ -5,7 +5,7 @@ namespace DroneStrikers.Combat
     public class SimpleBodyDamageSource : MonoBehaviour, IDamageSource
     {
         [SerializeField] private int _contactDamage = 1;
-        public int ContactDamage => _contactDamage;
+        public float ContactDamage => _contactDamage;
 
         public IDestructionContextReceiver InstigatorContextReceiver => null;
     }
